refactored code

+++ src/main/java/com/example/authdemo/services/TargetJsonService.java
@@ -1,49 +1,37 @@
 package com.example.authdemo.services;
 
 import com.example.authdemo.repository.TargetFileRepository;
-import com.example.authdemo.repository.entities.FileEntity;
-import com.example.authdemo.repository.entities.SourceFileEntity;
 import com.example.authdemo.repository.entities.TargetFileEntity;
-import com.example.authdemo.util.TargetCSVHelper;
+import com.example.authdemo.util.TargetJsonHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
 @Transactional
-public class TargetCSVService {
+public class TargetJsonService {
 
     private final TargetFileRepository repository;
 
     @Autowired
-    public TargetCSVService(TargetFileRepository repository) {
+    public TargetJsonService(TargetFileRepository repository) {
         this.repository = repository;
     }
 
     public void save(String username, MultipartFile file) {
         try {
-            List<TargetFileEntity> targetFileEntities = TargetCSVHelper.csvToTargetFile(username, file.getInputStream());
+            List<TargetFileEntity> targetFileEntities = TargetJsonHelper.jsonToTargetFile(username, file.getInputStream());
+            if (targetFileEntities == null)
+                throw new IOException("fail to store json data: ");
             repository.deleteByUsername(username);
             repository.saveAll(targetFileEntities);
         } catch (IOException e) {
-            throw new RuntimeException("fail to store csv data: " + e.getMessage());
+            throw new RuntimeException("fail to store json data: " + e.getMessage());
         }
-    }
-
-    public List<FileEntity> getFileEntities(String username) {
-        List<TargetFileEntity> sourceFileEntities = new ArrayList<>(this.getAll(username));
-        List<FileEntity> resultant = new ArrayList<>();
-        sourceFileEntities.forEach(fileEntity -> {
-            if (fileEntity.getSourceFileEntity() == null) {
-                resultant.add(fileEntity);
-            }
-        });
-        return resultant;
     }
 
     public List<TargetFileEntity> getAll(String username) {
